refactor: move titlebar leading inset support
