+++ Sources/Windowing/TitlebarLeadingInsetSupport.swift
@@ -1,0 +1,35 @@
+import AppKit
+import SwiftUI
+
+/// Reads the leading inset required to clear traffic lights + left titlebar accessories.
+final class TitlebarLeadingInsetPassthroughView: NSView {
+    override var mouseDownCanMoveWindow: Bool { false }
+    override func hitTest(_ point: NSPoint) -> NSView? { nil }
+}
+
+struct TitlebarLeadingInsetReader: NSViewRepresentable {
+    @Binding var inset: CGFloat
+
+    func makeNSView(context: Context) -> NSView {
+        let view = TitlebarLeadingInsetPassthroughView()
+        view.setFrameSize(.zero)
+        return view
+    }
+
+    func updateNSView(_ nsView: NSView, context: Context) {
+        DispatchQueue.main.async {
+            guard let window = nsView.window else { return }
+            // Start past the traffic lights
+            var leading: CGFloat = 78
+            // Add width of all left-aligned titlebar accessories
+            for accessory in window.titlebarAccessoryViewControllers
+                where accessory.layoutAttribute == .leading || accessory.layoutAttribute == .left {
+                leading += accessory.view.frame.width
+            }
+            leading += 0
+            if leading != inset {
+                inset = leading
+            }
+        }
+    }
+}
